refactor(metadata_service): modularize configuration, schemas, and dynamic auth
- Move environment loading and constants to config.py
- Move RegulationDocumentMetadata schema to schemas/metadata.py
- Move DynamicBearerAuth class to auth/http_auth.py
- Remove internal main_test and __main__ block (moved to scripts/example_metadata_service.py)
- Simplify metadata service initialization by using imported modules
